requirements.txt
bitarray==3.6.1
bitstring==4.3.1
bleak==1.1.0
contourpy==1.3.3
cycler==0.12.1
enum-compat==0.0.3
fonttools==4.59.1
joblib==1.5.1
kiwisolver==1.4.9
matplotlib==3.10.5
muselsl==2.3.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pexpect==4.9.0
pillow==11.3.0
ptyprocess==0.7.0
pygatt==5.0.0
pylsl==1.17.6
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
pyimgui==2.0.0
pygame==2.5.2
